backend/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
supabase>=2.0
python-dotenv>=1.0
